Add the ability to determine units without formatting
This PR adds a new bytes unit called general for the lack of a better term, but which is the intersection of
both the binary byte (1024 MiB = 1 GiB) format and the decimal byte (1000 MB = 1 GB) format, using the binary
scale and decimal units (e.g., 1024 MB = 1 GB). The format string marker is 'bd' as it's the combo of the two
existing markers; I thought about making it 'c' (because it's between 'b' and 'd'), but it does not make any
sense.

It also simplifies the format method by using shared code between the existing two formats, as well as
slightly simplifying the approach to determining the unit suffix. This will also print properly the Yotta scale
when the value _exeeds_ every unit (e.g., 1024 YiB or 1000 YB).

There is one downside to adding the third format string, which is that the unformat function will not properly
recognize the units (MB in binary-decimal uses a different scale from MB in decimal). Developers that use the
format with the new marker should simply avoid using the unformat string with it. For what it's worth, I expect
that the unformat method does not do what most people expect because they probably want MB to be 1024 KB
("binary" bytes) rather than the Mac-style 1000 KB ("decimal" bytes).