Enable setting language and format parameters on a Numbro instance
Store language, zeroFormat, defaultFormat abd defaultCurrencyFormat in
Numbro._config and prefer their value, if available, to global ones in
formatting methods.

Add the language method to the Numbro prototype for setting instance language
and getting inctance language, if available, or the global one.

Add setLanguage, zeroFormat, defaultFormat and defaultCurrencyFormat to the
Numbro prototype to set the instance configuration.

Introduce culture and seCulture methods to numbro instances to follow the
global methods and deprecate language and setLanguage.

Make both setLanguage and setCulture use the same search algorithm, if the
exact culture was not matched.  First try the language prefix, then try the
country suffix.  Preferring the language prefix is usual and once setLanguage
is gone, setCulture shoudl support it.

Do not call unformat from the prototype from the Numbro constructor; instance
variables would not be accessible.

Add tests for the getting  and setting the inxtance configuration.

Increase the maximum argument count to 5 in jshint checks.