Enable multiple numbro factory functions with separate configurations
Add a clone method returning another numbro factory function, which maintains
its own configuration, independent of the source one or of the global one.

Prevent modifying the source factory configuration from numbro instances
created by it.

Add tests for configuring the clone and creating instances with it.

Introduce culture and seCulture methods to numbro instances to follow the
global methods and deprecate language and setLanguage.

Make both setLanguage and setCulture use the same search algorithm, if the
exact culture was not matched.  First try the language prefix, then try the
country suffix.  Preferring the language prefix is usual and once setLanguage
is gone, setCulture shoudl support it.