Fix #190 Allow abbreviation precision for < 4
Currently for some reason when using abbreviations like `'0,0a`  you
can't specify the number of visible digits of precision for any value
less than 4. So while `'4,0a'` will change `1234567` to `1,234k` the
result of `'3,0a'` is `1m` instead of `1.23m` as expected.

This is primarily caused by a conditional that skips setting the
precision in the formatting.

This commit changes how the algorithm determines if decimal places will
be required and how many are required. We first compute the abbreviated
unitized value and then use the length of that value to determine how
many decimal places will be required.