Optimize repeated calls to Math.pow(10, x) with a cache
There are a number of places where a fixed calculation of Math.pow is
used. Typically to determine units so 10^3, 10^6, 10^9 etc.

These are being called multiple times for every format call.
Additionally formatting could potentially be triggered quite frequently
if it is part of any data binding or observables.

I felt it made sense to optimize these by caching the values after the
first call.

To exclude a potential memory leak edge case the cache is bounded to
100 which covers number lengths up to 100.