en-IN localization
Added support for customizable grouping widths. In the delimiters, specify group: [2,3], then format() will group like this

12,00,00,000

To make this work, we build up the grouping regex according to the group configuration. For en-IN, we get this:

/(\d)(?=(?:\d{2})*(?:\d{3})(?!\d))/

It matches a single digit, followed by non capturing look-ahead that matches zero or more 2 digit groups, followed by a 3 digit group.

The default for the group property is [3] which matches the current behaviour of grouping by thousands.

It should be possible to pass [4] and get grouping by ten thousands for e.g. Japanese. I'm not aware of more complex groupings.