Fix Hebrew words for million and billion
The word "million" was misspelled: it had a missing letter.
It makes it longish, but there's no common way to abbreviate
it in Hebrew, and the words in Thai, Persian, and some other
languages are longish, too, so it's not a problem.

The word "billion" was misspelled, too, but more importantly,
it's not actually used in Hebrew. Both in the colloquial language
and in the standard published by the Academy of the Hebrew Language,
the recommended word for one with nine zeros is "milliard".
Reference: Academy meeting 329, January 21, 2013
link: https://hebrew-academy.org.il/wp-content/uploads/meeting-329.pdf
I changed it and added a test for it.

Also, add tests for "trillion".